changed type from float to double

+++ src/geek-calculator/Function.hpp
@@ -12,11 +12,11 @@ public:
 
     Function(const QString& str_func, const int number_of_variables=1);
     ~Function();
-    float get_value();
-    float oneVariableIntegral(const float a, const float b);
-    float twoVariableIntegral(const float a1, const float b1, const float a2, const float b2);
-    float firstDerivative(const float point);
-    float secondDerivative(const float point);
+    double get_value();
+    double oneVariableIntegral(const double a, const double b);
+    double twoVariableIntegral(const double a1, const double b1, const double a2, const double b2);
+    double firstDerivative(const double point);
+    double secondDerivative(const double point);
     QString get_string_function() const;
 
     
@@ -24,10 +24,10 @@ public:
 
 private:
     QString m_function;
-    float function_value;
+    double function_value;
     int m_number_of_variables;
     bool check_function();
-    std::vector<float> linspace(const float a, const float b, const size_t n);
+    std::vector<double> linspace(const double a, const double b, const size_t n);
